requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.6
bitcoin==1.1.42
black==21.4b1
blockcypher==1.0.91
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
django-countries==7.1
django-cryptocurrency-payment==0.1.1
django-phonenumber-field==5.1.0
forex-python==1.5
idna==2.10
merchant-wallet==0.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumbers==8.12.22
Pillow==8.2.0
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
